+++ src/admin/routes/attributes/set/[id]/components/index.ts
@@ -1,0 +1,2 @@
+export * from './AttributeSetAttributesSection'
+export * from './AttributeSetGeneralSection'