+++ src/admin/types/global.d.ts
@@ -1,0 +1,3 @@
+declare const __BACKEND_URL__: string
+declare const __BASE__: string
+declare const __STOREFRONT_URL__: string