+++ src/api/admin/plugin/attribute-set/route.ts
@@ -1,7 +1,10 @@
 import { MedusaRequest, MedusaResponse, refetchEntity } from "@medusajs/framework";
 
-import { AdminCreateAttributeSetType } from "./validators";
+import { AdminCreateAttributeSetType, AdminGetAttributesSetsParamsType } from "./validators";
 import { createAttributeSetWorkflow } from "../../../../workflows/attribute-set/workflows/create-attribute-set";
+import { ContainerRegistrationKeys } from "@medusajs/framework/utils";
+import { AttributeSet } from "../../../../types/attribute";
+import { PaginatedResponse } from "@medusajs/framework/types";
 
 export const POST = async (req: MedusaRequest<AdminCreateAttributeSetType>, res: MedusaResponse) => {
     const { result: [attributeSet] } = await createAttributeSetWorkflow(req.scope).run({
@@ -16,4 +19,16 @@ export const POST = async (req: MedusaRequest<AdminCreateAttributeSetType>, res:
     )
     
     return res.status(201).json({ attribute_set: response })
+}
+
+export const GET = async (req:MedusaRequest<AdminGetAttributesSetsParamsType>, res: MedusaResponse) => {
+    const query = req.scope.resolve(ContainerRegistrationKeys.QUERY)
+
+    const { data: attributeSets, metadata } = await query.graph({
+        entity: 'attribute_set',
+        filters: req.filterableFields,
+        ...req.queryConfig
+    })
+
+    return res.status(200).json({ attributeSets, count: metadata?.count, offset: metadata?.skip, limit: metadata?.take })
 }