+++ src/types/attribute/http/attribute-set/index.ts
@@ -1,0 +1,10 @@
+import { Attribute } from "../attribute"
+
+export type AttributeSet = {
+    id: string
+    name: string
+    description?: string
+    handle: string
+    metadata: Record<string, unknown> | null
+    attributes: Attribute[]
+}