+++ src/workflows/attribute-set/steps/index.ts
@@ -1,1 +1,3 @@
 export * from './create-attribute-set'
+export * from './batch-link-attribute-set-attributes'
+export * from './update-attribute-set'