+++ src/workflows/attribute/workflows/delete-attribute.ts
@@ -1,0 +1,15 @@
+import { createWorkflow, WorkflowResponse } from "@medusajs/framework/workflows-sdk"
+import { deleteAttributeStep } from "../steps"
+
+export const deleteAttributeWorkflowId = 'delete-attribute-worklfow'
+
+type DeleteAttributeWorkflowInput = {
+    id: string
+}
+
+export const deleteAttributeWorkflow = createWorkflow(
+    deleteAttributeWorkflowId,
+    (input: DeleteAttributeWorkflowInput) => {
+        return new WorkflowResponse(deleteAttributeStep(input))
+    }
+)