+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+- UI Routes in admin panel to manage attributes, attribute sets and possible values
+- Added ui_component to attribute schema, to allow admin widgets in product to understand how they should ask for the value input
+- API Routes and workflows to facilitate interacting with the UI components 
+
 ## 2.0.2 - 2025-06-10
 ### Fixed
 - Updating an attribute value, with same value as other existing ones, causes deletion of the other values, due to a bug in the `productsUpdatedHookHandler`